+++ core/inc/kexecutive.h
@@ -4,7 +4,7 @@
 /** RK0 - The Embedded Real-Time Kernel '0'                                   */
 /** (C) 2026 Antonio Giacomelli <dev@kernel0.org>                             */
 /**                                                                           */
-/** VERSION: V0.30.0                                                           */
+/** VERSION: V0.40.0                                                           */
 /**                                                                           */
 /** You may obtain a copy of the License at :                                 */
 /** http://www.apache.org/licenses/LICENSE-2.0                                */
@@ -33,7 +33,7 @@ extern "C" {
 #include <ksema.h>
 #include <kmutex.h>
 #include <kmesgq.h>
-#include <kexchange.h>
+#include <krendezvous.h>
 #include <kmrm.h>
 #include <ktimer.h>
 #include <ktrace.h>