+++ core/src/kmutex.c
@@ -4,7 +4,7 @@
 /** RK0 - The Embedded Real-Time Kernel '0'                                   */
 /** (C) 2026 Antonio Giacomelli <dev@kernel0.org>                             */
 /**                                                                           */
-/** VERSION: V0.30.0                                                           */
+/** VERSION: V0.40.0                                                           */
 /**                                                                           */
 /** You may obtain a copy of the License at :                                 */
 /** http://www.apache.org/licenses/LICENSE-2.0                                */
@@ -117,6 +117,7 @@ static inline void kMutexUpdateOwnerPrio_(struct RK_OBJ_TCB *ownerTcb)
             K_ASSERT(err == RK_ERR_SUCCESS);
 
             currTcbPtr->priority = newPrio;
+            kTraceRecordTaskPrio(currTcbPtr, oldPrio, newPrio);
 
             err =
                 kTCBQEnq(
@@ -128,7 +129,9 @@ static inline void kMutexUpdateOwnerPrio_(struct RK_OBJ_TCB *ownerTcb)
         }
         else
         {
+            RK_PRIO oldPrio = currTcbPtr->priority;
             currTcbPtr->priority = newPrio;
+            kTraceRecordTaskPrio(currTcbPtr, oldPrio, newPrio);
         }
 
         RK_BARRIER